Copied Back to Basics commit in EDAcraft_v7 branch for single direction movement debugging
